Enhancing per radio user space rate control
To facilitate per radio configurations, the AccessPoint is initialized
with the config dict. RateMan object is therefore agnostic of the
radio configurations within the AP object. RateMan applies the system
and rate control config. If user space rate control algorithm (rc_alg) is to be
used, then one asyncio task per radio is created. A rc_alg object can
create subtasks if necessary as the loop is passed to it. In the further
development steps, the task management can be improved such that the
main rateman object takes care of maintaining a list of all tasks and
cancelling these tasks upon completion/interruption.