accesspoint: remove misleading supported_rates property
AccessPoint objects had a property called supported_rates, which would return
a dictionary of the entire MCS rate space covered by ORCA. This was misleading,
because it suggested that the set was restricted to the rates that the AP
actually supports, which may not be the case. I've replaced it with
a new function called get_rate_info(), which will provide information on a given
rate index, such as bandwidth, SGI, number of spatial streams, etc.

Signed-off-by: Arne Kappen <arne.kappen@hhi.fraunhofer.de>