fix parsing of TPC capabilities
This fixes parsing of the TPC ranges block(s) which lead to issues with
mt7615 chips.

The list of available power levels was constructed using the values
'start_pwr' and 'pwr_width'. For mt7615, which defines its start_pwr as
-16, this produces the list [-8.0, -7.5, -7.0, ..., 7.5, 8.0] which
contains the real dBm values. However, the API only uses the power
indices and thus, the values from the list above cannot be used to set
the TX power. Additionally, the abovementioned list used 'n_indices + 1'
which produced one excess value which is outside the range.

With the fix, the list of available power levels for mt7615 is now
generated as [0, 1, 2, 3, 4, ..., 29, 30, 31].

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>