add missing check of return value in parsing
The function parse_str() uses the subrouting next_field() to determine
the length of the string between the current position and the next
separator. This length is then used for a simple memcpy to copy the
string into a separate buffer.
parse_str() checks if the length of the string exceeds the size of the
buffer to prevent buffer overflow. However, next_field() may also return
the value -1 in case strchr() returns NULL. This condition hasn't been
covered in the code.

In case an broken trace is parsed, memcpy() is called with the value -1
for its argument 'size'. Since the type of the argument 'size' is
size_t (which is an unsigned type), -1 is implicitly casted and results
in the maximum value of size_t. Attempting to memcpy with this value
causes a SEGFAULT.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>